dialogs unified into home page
