+++ frontend/src/components/Dialog/Rom/SearchRom.vue
@@ -1,103 +1,88 @@
 <script setup lang="ts">
-import type { SearchRomSchema } from "@/__generated__";
+import type { RomSchema } from "@/__generated__";
+import PlatformIcon from "@/components/Platform/PlatformIcon.vue";
 import romApi from "@/services/api/rom";
-import storeRoms, { type Rom } from "@/stores/roms";
+import storeGalleryView from "@/stores/galleryView";
 import type { Events } from "@/types/emitter";
+import { languageToEmoji, regionToEmoji } from "@/utils";
+import { identity, isNull } from "lodash";
 import type { Emitter } from "mitt";
 import { inject, onBeforeUnmount, ref } from "vue";
+import { useRouter } from "vue-router";
 import { useDisplay, useTheme } from "vuetify";
 
-const { xs, mdAndDown, lgAndUp } = useDisplay();
-const show = ref(false);
-const rom = ref<Rom | null>(null);
-const romsStore = storeRoms();
-const renameAsIGDB = ref(false);
-const searching = ref(false);
-const searchTerm = ref("");
-const searchBy = ref("Name");
-const searchExtended = ref(false);
-const matchedRoms = ref<SearchRomSchema[]>([]);
 const theme = useTheme();
+const { xs, mdAndDown, lgAndUp } = useDisplay();
+const galleryViewStore = storeGalleryView();
+const show = ref(false);
+const searching = ref(false);
+const router = useRouter();
+const searchedRoms = ref();
+const filteredRoms = ref();
+const platforms = ref();
+const selectedPlatform = ref();
+const searchValue = ref("");
+const showRegions = isNull(localStorage.getItem("settings.showRegions"))
+  ? true
+  : localStorage.getItem("settings.showRegions") === "true";
+const showLanguages = isNull(localStorage.getItem("settings.showLanguages"))
+  ? true
+  : localStorage.getItem("settings.showLanguages") === "true";
+const showSiblings = isNull(localStorage.getItem("settings.showSiblings"))
+  ? true
+  : localStorage.getItem("settings.showSiblings") === "true";
+
 const emitter = inject<Emitter<Events>>("emitter");
-emitter?.on("showSearchRomDialog", (romToSearch) => {
-  rom.value = romToSearch;
-  searchTerm.value = romToSearch.name || romToSearch.file_name_no_tags || "";
+emitter?.on("showSearchRomDialog", () => {
   show.value = true;
-  searchRom();
 });
 
-// Functions
-function toggleExtended() {
-  searchExtended.value = !searchExtended.value;
+function clearFilter() {
+  selectedPlatform.value = null;
 }
 
-async function searchRom() {
-  if (!rom.value) return;
-
+async function searchRoms() {
   // Auto hide android keyboard
   const inputElement = document.getElementById("search-text-field");
   inputElement?.blur();
+  searching.value = true;
+  searchedRoms.value = (
+    await romApi.getRoms({ searchTerm: searchValue.value, size: 250 })
+  ).data.items.sort((a, b) => {
+    return a.platform_name.localeCompare(b.platform_name);
+  });
+  platforms.value = [
+    ...new Set(
+      searchedRoms.value.map(
+        (rom: { platform_name: string }) => rom.platform_name
+      )
+    ),
+  ];
+  filterRoms();
+  searching.value = false;
+}
 
-  if (!searching.value) {
-    searching.value = true;
-    await romApi
-      .searchRom({
-        romId: rom.value.id,
-        searchTerm: searchTerm.value,
-        searchBy: searchBy.value,
-        searchExtended: searchExtended.value,
-      })
-      .then((response) => {
-        matchedRoms.value = response.data;
-      })
-      .catch((error) => {
-        emitter?.emit("snackbarShow", {
-          msg: error.response.data.detail,
-          icon: "mdi-close-circle",
-          color: "red",
-        });
-      })
-      .finally(() => {
-        searching.value = false;
-      });
+async function filterRoms() {
+  if (!selectedPlatform.value) {
+    filteredRoms.value = searchedRoms.value;
+  } else {
+    filteredRoms.value = searchedRoms.value.filter(
+      (rom: { platform_name: string }) =>
+        rom.platform_name == selectedPlatform.value
+    );
   }
 }
 
-async function updateRom(matchedRom: SearchRomSchema) {
-  if (!rom.value) return;
-
-  show.value = false;
-  emitter?.emit("showLoadingDialog", { loading: true, scrim: true });
-
-  Object.assign(rom.value, matchedRom);
-
-  await romApi
-    .updateRom({ rom: rom.value, renameAsIGDB: renameAsIGDB.value })
-    .then(({ data }) => {
-      emitter?.emit("snackbarShow", {
-        msg: "Rom updated successfully!",
-        icon: "mdi-check-bold",
-        color: "green",
-      });
-      romsStore.update(data);
-      emitter?.emit("refreshView", null);
-    })
-    .catch((error) => {
-      emitter?.emit("snackbarShow", {
-        msg: error.response.data.detail,
-        icon: "mdi-close-circle",
-        color: "red",
-      });
-    })
-    .finally(() => {
-      emitter?.emit("showLoadingDialog", { loading: false, scrim: false });
-    });
+function romDetails(rom: RomSchema) {
+  router.push({
+    name: "rom",
+    params: { rom: rom.id },
+  });
+  closeDialog();
 }
 
 function closeDialog() {
   show.value = false;
-  searchBy.value = "Name";
-  searchExtended.value = false;
 }
 
 onBeforeUnmount(() => {
@@ -126,13 +111,15 @@ onBeforeUnmount(() => {
     >
       <v-toolbar density="compact" class="bg-terciary">
         <v-row class="align-center" no-gutters>
-          <v-col cols="9" xs="9" sm="10" md="10" lg="11">
-            <v-icon icon="mdi-search-web" class="ml-5" />
+          <v-col cols="10" xs="10" sm="11" md="11" lg="11">
+            <v-icon icon="mdi-magnify" class="ml-5" />
+            <v-avatar :rounded="0" :size="30" class="mx-4"
+              ><v-img src="/assets/isotipo.svg"
+            /></v-avatar>
           </v-col>
-          <v-col>
+          <v-col cols="2" xs="2" sm="1" md="1" lg="1">
             <v-btn
               @click="closeDialog"
-              class="bg-terciary"
               rounded="0"
               variant="text"
               icon="mdi-close"
@@ -146,54 +133,40 @@ onBeforeUnmount(() => {
 
       <v-toolbar density="compact" class="bg-primary">
         <v-row class="align-center" no-gutters>
-          <v-col cols="5" xs="5" sm="5" md="6" lg="8">
+          <v-col cols="6" xs="6" sm="6" md="6" lg="7">
             <v-text-field
+              autofocus
               id="search-text-field"
-              @keyup.enter="searchRom()"
-              @click:clear="searchTerm = ''"
-              class="bg-terciary"
-              v-model="searchTerm"
-              label="search"
+              @keyup.enter="searchRoms"
+              @click:clear="searchRoms"
+              v-model="searchValue"
+              label="Search"
               hide-details
+              class="bg-terciary"
               clearable
             />
           </v-col>
-          <v-col cols="3" xs="3" sm="3" md="2" lg="2">
+          <v-col cols="4" xs="4" sm="5" md="5" lg="4">
             <v-select
-              label="by"
+              @click:clear="clearFilter"
+              clearable
+              label="Platform"
               class="bg-terciary"
-              :items="['ID', 'Name']"
-              v-model="searchBy"
               hide-details
-            />
+              v-model="selectedPlatform"
+              @update:model-value="filterRoms"
+              :items="platforms"
+            >
+            </v-select>
           </v-col>
-
-          <v-col cols="2" xs="2" sm="2" md="2" lg="1">
-            <v-tooltip
-              location="top"
-              class="tooltip"
-              transition="fade-transition"
-              text="Extended search to match by alternative names. This will take longer."
-              open-delay="500"
-              ><template v-slot:activator="{ props }">
-                <v-btn
-                  v-bind="props"
-                  @click="toggleExtended"
-                  class="bg-terciary"
-                  :color="searchExtended ? 'romm-accent-1' : ''"
-                  rounded="0"
-                  variant="tonal"
-                  icon="mdi-layers-search-outline"
-                  block /></template></v-tooltip
-          ></v-col>
-          <v-col cols="2" xs="2" sm="2" md="2" lg="1">
+          <v-col cols="2" xs="2" sm="1" md="1" lg="1">
             <v-btn
               type="submit"
-              @click="searchRom()"
+              @click="searchRoms"
               class="bg-terciary"
               rounded="0"
               variant="text"
-              icon="mdi-search-web"
+              icon="mdi-magnify"
               block
               :disabled="searching"
             />
@@ -218,7 +191,7 @@ onBeforeUnmount(() => {
         </v-row>
         <v-row
           class="justify-center align-center loader-searching fill-height"
-          v-show="!searching && matchedRoms.length == 0"
+          v-show="!searching && searchedRoms?.length == 0"
           no-gutters
         >
           <span>No results found</span>
@@ -227,16 +200,16 @@ onBeforeUnmount(() => {
           <v-col
             class="pa-1"
             cols="4"
-            xs="4"
-            sm="3"
-            md="3"
+            sm="2"
+            md="2"
             lg="2"
+            xl="2"
             v-show="!searching"
-            v-for="matchedRom in matchedRoms"
+            v-for="rom in filteredRoms"
           >
             <v-hover v-slot="{ isHovering, props }">
               <v-card
-                @click="updateRom(matchedRom)"
+                @click="romDetails(rom)"
                 v-bind="props"
                 class="matched-rom"
                 :class="{ 'on-hover': isHovering }"
@@ -246,14 +219,18 @@ onBeforeUnmount(() => {
                   <v-img
                     v-bind="props"
                     :src="
-                      !matchedRom.url_cover
+                      !rom.igdb_id && !rom.moby_id && !rom.has_cover
+                        ? `/assets/default/cover/big_${theme.global.name.value}_unmatched.png`
+                        : !rom.has_cover
                         ? `/assets/default/cover/big_${theme.global.name.value}_missing_cover.png`
-                        : matchedRom.url_cover
+                        : `/assets/romm/resources/${rom.path_cover_l}`
                     "
                     :lazy-src="
-                      !matchedRom.url_cover
+                      !rom.igdb_id && !rom.moby_id && !rom.has_cover
+                        ? `/assets/default/cover/small_${theme.global.name.value}_unmatched.png`
+                        : !rom.has_cover
                         ? `/assets/default/cover/small_${theme.global.name.value}_missing_cover.png`
-                        : matchedRom.url_cover
+                        : `/assets/romm/resources/${rom.path_cover_s}`
                     "
                     :aspect-ratio="3 / 4"
                   >
@@ -270,58 +247,78 @@ onBeforeUnmount(() => {
                     </template>
                     <v-expand-transition>
                       <div
-                        v-if="isHovering || !matchedRom.url_cover"
+                        v-if="isHovering || !rom.has_cover"
                         class="translucent text-caption"
+                        :class="{
+                          'text-truncate':
+                            galleryViewStore.current == 0 && !isHovering,
+                        }"
                       >
-                        <v-list-item>{{ matchedRom.name }}</v-list-item>
+                        <v-list-item>{{ rom.name }}</v-list-item>
                       </div>
                     </v-expand-transition>
-                    <v-row no-gutters class="text-white pa-1">
-                      <v-tooltip
-                        location="top"
-                        class="tooltip"
-                        transition="fade-transition"
-                        text="IGDB matched"
-                        open-delay="500"
-                        ><template v-slot:activator="{ props }">
-                          <v-avatar
-                            v-bind="props"
-                            v-if="matchedRom.igdb_id"
-                            class="mr-1"
-                            size="30"
-                            rounded="1"
-                          >
-                            <v-img
-                              src="/assets/scrappers/igdb.png"
-                            /> </v-avatar></template
-                      ></v-tooltip>
-                      <v-tooltip
-                        location="top"
-                        class="tooltip"
-                        transition="fade-transition"
-                        text="Mobygames matched"
-                        open-delay="500"
-                        ><template v-slot:activator="{ props }">
-                          <v-avatar
-                            v-bind="props"
-                            v-if="matchedRom.moby_id"
-                            class="mr-1"
-                            size="30"
-                            rounded="1"
-                          >
-                            <v-img
-                              src="/assets/scrappers/moby.png"
-                            /> </v-avatar></template
-                      ></v-tooltip>
-                    </v-row> </v-img
-                ></v-hover>
+                    <v-row no-gutters class="text-white px-1">
+                      <v-chip
+                        v-if="
+                          rom.regions.filter(identity).length > 0 && showRegions
+                        "
+                        :title="`Regions: ${rom.regions.join(', ')}`"
+                        class="translucent mr-1 mt-1 px-1"
+                        :class="{ 'emoji-collection': rom.regions.length > 3 }"
+                        density="compact"
+                      >
+                        <span
+                          class="emoji"
+                          v-for="region in rom.regions.slice(0, 3)"
+                        >
+                          {{ regionToEmoji(region) }}
+                        </span>
+                      </v-chip>
+                      <v-chip
+                        v-if="
+                          rom.languages.filter(identity).length > 0 &&
+                          showLanguages
+                        "
+                        :title="`Languages: ${rom.languages.join(', ')}`"
+                        class="translucent mr-1 mt-1 px-1"
+                        :class="{
+                          'emoji-collection': rom.languages.length > 3,
+                        }"
+                        density="compact"
+                      >
+                        <span
+                          class="emoji"
+                          v-for="language in rom.languages.slice(0, 3)"
+                        >
+                          {{ languageToEmoji(language) }}
+                        </span>
+                      </v-chip>
+                      <v-chip
+                        v-if="
+                          rom.siblings &&
+                          rom.siblings.length > 0 &&
+                          showSiblings
+                        "
+                        :title="`${rom.siblings.length + 1} versions`"
+                        class="translucent mr-1 mt-1"
+                        density="compact"
+                      >
+                        +{{ rom.siblings.length }}
+                      </v-chip>
+                    </v-row></v-img
+                  ></v-hover
+                >
                 <v-card-text>
                   <v-row class="pa-1 align-center">
                     <v-col class="pa-0 ml-1 text-truncate">
-                      <span :title="matchedRom.name">{{
-                        matchedRom.name
-                      }}</span>
+                      <span>{{ rom.name }}</span>
                     </v-col>
+                    <v-avatar :rounded="0" size="20" class="ml-2">
+                      <platform-icon
+                        :key="rom.platform_slug"
+                        :slug="rom.platform_slug"
+                      />
+                    </v-avatar>
                   </v-row>
                 </v-card-text>
               </v-card>
@@ -329,22 +326,15 @@ onBeforeUnmount(() => {
           </v-col>
         </v-row>
       </v-card-text>
-
-      <v-divider class="border-opacity-25" :thickness="1" />
-
-      <v-toolbar class="bg-terciary" density="compact">
-        <v-checkbox
-          v-model="renameAsIGDB"
-          label="Rename rom"
-          class="ml-3"
-          hide-details
-        />
-      </v-toolbar>
     </v-card>
   </v-dialog>
 </template>
 
 <style scoped>
+.tooltip :deep(.v-overlay__content) {
+  background: rgba(201, 201, 201, 0.98) !important;
+  color: rgb(41, 41, 41) !important;
+}
 .scroll {
   overflow-y: scroll;
 }
@@ -376,8 +366,12 @@ onBeforeUnmount(() => {
   backdrop-filter: blur(10px);
   text-shadow: 1px 1px 1px #000000, 0 0 1px #000000;
 }
-.tooltip :deep(.v-overlay__content) {
-  background: rgba(201, 201, 201, 0.98) !important;
-  color: rgb(41, 41, 41) !important;
+
+.emoji-collection {
+  mask-image: linear-gradient(to right, black 0%, black 70%, transparent 100%);
+}
+
+.emoji {
+  margin: 0 2px;
 }
 </style>
